Update thread naming to separate cmd
