+++ zns.trace/util/helpers.py
@@ -1,13 +1,46 @@
 #! /usr/bin/python3
 
+def init_tid_map(tid_map):
+    tid_map["nvme_cmd_write"] = 0;
+    tid_map["nvme_cmd_read"] = 1;
+    tid_map["nvme_zone_reset"] = 2;
+    tid_map["compaction"] = 3;
+    tid_map["flush"] = 4;
+    tid_map["f2fs_move_data"] = 5;
+    tid_map["f2fs_submit_page_write"] = 6;
+    tid_map["vfs_open"] = 7;
+    tid_map["vfs_create"] = 8;
+    tid_map["vfs_fsync"] = 9;
+    tid_map["vfs_rename"] = 10;
+    tid_map["vfs_unlink"] = 11;
+    tid_map["fcntl_set_rw_hint"] = 12;
+    tid_map["mm_do_writepages"] = 13;
+
+# 0 is lowest process in timeline, otherwise in increasing order
+def get_pid(name):
+    if 'nvme' in name:
+        return 0
+    elif 'compaction' in name or 'flush' in name:
+        return 1
+    elif 'vfs' in name or 'rw_hint' in name:
+        return 2
+    elif 'f2fs' in name:
+        return 3
+    elif 'mm' in name:
+        return 4
+
 def get_cmd(cmd):
-    match cmd:
-        case 0x1:
+    match str(cmd):
+        case "1":
             return "nvme_cmd_write"
-        case 0x2:
+        case "2":
             return "nvme_cmd_write"
-        case 0x7d:
+        case "15":
+            return "nvme_zone_reset" # set for a zone reset from nvme_cmd_resv_release
+        case "7d":
             return "nvme_cmd_zone_append"
+        case _:
+            return None
 
 def get_hint(hint):
     match hint:
